ENH Metadata propagates into ndarray or scalar
Metadata provided by the subclassed reader will be propagated into an
ndarray, so that it can be indexed in the same way as the planes in the
frame. When metadata is provided incomplete, a warning message is
displayed and that field is not propagated.

When all fields of the planes are equal, the field is converted to a
scalar.