Add reader vor SPE files
Apart from the actual image data, also some metadata from the file header is
read and stored in a dict named "metadata". Not all metadata fields are
currently read from a file (but reading of additional fields is very easy to
add). Reading XML metadata as per SPE 3.0 specification is not supported.