FIX Plots to frame pad mode
This fixes an issue in python 3. `np.pad` takes a mode, that is a
string, which should be unicode in py3 and bytes in py2.7. Currently it
failed in python 3, and together with `from __future__ import
unicode_literals` it needs to be converted in Py2.7. The solution is a
`str()` operation.