Use availability functions to check tests to skip
Instead of trying to deduce what functionality may or may not be present
by using imports, simply use the availability function to determine if
the functionality is present or not. This way we can rely on whatever is
the best strategy for that module when running the tests for it. Also
this gives us the opportunity to verify that the availability functions
are working as intended.