Add dependency on imageio; remove matplotlib-based reader.
matplotlib's imread is now anyways a thin shim around pillow (a
dependency of matplotlib), so if you have matplotlib installed, you may
as well use imageio's imread: imageio is a smaller dependency, and has
more consistent behavior (it doesn't return pngs as floats).

imageio is also a dependency of scikit-image (since 0.15), so it seems
reasonable to directly depend on it instead.  For now, the scikit-image
reader was kept, as removing it (possibly in the future) would entail
removing support for the `plugin` kwarg, and thus an API break.