Don't emit warnings that duplicate exceptions on opening failure.
The error messages will already be displayed in the exception, so no
need to emit them as warnings too.  Moreover, warnings are somewhat more
annoying to suppress than exceptions.

(One *could* consider still emitting the warnings if e.g. the first
reader failed but the second one succeeds, but I'm not sure it's worth
the trouble.)